Clean: move ft_*_e() func to separate files and directory
